+++ crates/nostr/CHANGELOG.md
@@ -69,6 +69,7 @@
 - Remove nip47 `multi_*` methods following spec cleanup (https://github.com/nostr-protocol/nips/pull/2210)
 - Remove `once_cell` dependency (https://github.com/rust-nostr/nostr/pull/1267)
 - Remove `kind` field in `CommentTarget::Coordinate` variant (https://github.com/rust-nostr/nostr/pull/1294)
+- Remove `Timestamp::as_u64` (https://github.com/rust-nostr/nostr/pull/1295)
 
 ### Fixed